requirements/dev.txt
-r base.txt

django-autofixture==0.8.0
django-reset==0.2.0
django-debug-toolbar==1.2.1
pudb==2014.1
coverage==3.6
django-discover-runner==1.0
django-extensions==1.3.8
django-nose==1.2
